Use CpuQuota instead of CpuShares (#69)
Using the CFS (Completely Fair Scheduler) with a period of 100 ms allows to
truly hard-limit the CPU usage of a container. Using `CpuShares` might have some
unintended consequences, such as containers being killed by Plancton because
they trespass the given CPU limit, which is possible when nothing else is
running on the system.